Including statistical attributes
Using `mathjs` to calculate stat results, although it's slower since its using the entire population rather than the frequency histogram. However, it's more readable and easy-to-use, which is important to the project now.

I may change this later if it becomes more important. I'll keep a dangling commit saved with the functions I made.